Week 4 Assignment: Functions
Completed the flowcharts, commented more lines of code and removed the
“- 1” from lottoType on lines 58 and 62. I made the change because I
realized the number to select within did not need to be decreased for
every number removed from the array since only the values were changed.
I also added code to prevent the user from breaking the program when an
improper value was entered.